+++ src/test/java/src/test/MessageTest.java
@@ -1,6 +1,8 @@
-package src;
+package src.test;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import src.Message;
+
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import static org.junit.jupiter.api.Assertions.*;
